feat: Add /-/reload http endpoint to trigger ACL reload
This implements the same interface that Prometheus offers for triggering
reloads at runtime, a http POST to /-/reload.  It will be very helpful
for Kubernetes deployments, where you may be able to edit the ACL file
at runtime, but there is generally not a good way of sending SIGHUP to
the daemon.  Instead, the typical pattern is a sidecar container to
watch for changes and make a http request to the main process.